// client-id/src/app/core/models/xuat-hang.ts
export class XuatHang {
    expId: string;
    expDate: Date;
    expNote: string;
    expTotal: number;
    cusId: string;
    cusName: string;
    userId: string;
    proId: string;
    proName: string;
    quantity: number;
    expPrice: number;

    constructor() {
        this.expId = null;
        this.expDate = null;
        this.expNote = null;
        this.expTotal = null;
        this.cusId = null;
        this.cusName = null;
        this.userId = null;
        this.proId = null;
        this.proName = null;
        this.quantity = null;
        this.expPrice = null;
    }
}